django/requirements.txt
django
tensorflow
djangorestframework
django-cors-headers
Pillow
